requirements.txt
beautifulsoup4==4.9.3
Pillow==8.2.0
moviepy==1.0.3
configparser==5.0.1
requests==2.23.0
selenium==3.141.0
urllib3==1.25.9